+++ test/torchaudio_unittest/backend/sox_io/smoke_test.py
@@ -4,6 +4,7 @@ import unittest
 
 from torchaudio.utils import sox_utils
 from torchaudio.backend import sox_io_backend
+from torchaudio._internal.module_utils import is_sox_available
 from parameterized import parameterized
 
 from torchaudio_unittest.common_utils import (
@@ -16,7 +17,7 @@ from .common import name_func
 
 
 skipIfNoMP3 = unittest.skipIf(
-    not sox_utils.is_sox_available() or
+    not is_sox_available() or
     'mp3' not in sox_utils.list_read_formats() or
     'mp3' not in sox_utils.list_write_formats(),
     '"sox_io" backend does not support MP3')
