+++ test/torchaudio_unittest/common_utils/case_utils.py
@@ -8,7 +8,10 @@ import unittest
 import torch
 from torch.testing._internal.common_utils import TestCase as PytorchTestCase
 import torchaudio
-from torchaudio._internal.module_utils import is_module_available
+from torchaudio._internal.module_utils import (
+    is_module_available,
+    is_sox_available
+)
 from torchaudio.utils import sox_utils
 
 from .backend_utils import set_audio_backend
@@ -77,7 +80,6 @@ class TestBaseMixin:
 
     def setUp(self):
         super().setUp()
-        # if self.backend != "sox_io" or sox_utils.is_sox_available():
         set_audio_backend(self.backend)
 
 
@@ -97,7 +99,7 @@ def skipIfNoModule(module, display_name=None):
 skipIfNoSoxBackend = unittest.skipIf(
     'sox' not in torchaudio.list_audio_backends(), 'Sox backend not available')
 skipIfNoCuda = unittest.skipIf(not torch.cuda.is_available(), reason='CUDA not available')
-skipIfNoSox = unittest.skipIf(not sox_utils.is_sox_available(), reason='Sox not available')
+skipIfNoSox = unittest.skipIf(not is_sox_available(), reason='Sox not available')
 
 
 def skipIfNoExtension(test_item):
