+++ torchaudio/csrc/pybind.cpp
@@ -3,8 +3,10 @@
 #ifdef INCLUDE_SOX
 #include <torchaudio/csrc/sox/effects.h>
 #include <torchaudio/csrc/sox/io.h>
+#endif
 
 PYBIND11_MODULE(_torchaudio, m) {
+#ifdef INCLUDE_SOX
   m.def(
       "get_info_fileobj",
       &torchaudio::sox_io::get_info_fileobj,
@@ -21,7 +23,5 @@ PYBIND11_MODULE(_torchaudio, m) {
       "apply_effects_fileobj",
       &torchaudio::sox_effects::apply_effects_fileobj,
       "Decode audio data from file-like obj and apply effects.");
-}
-#else
-PYBIND11_MODULE(_torchaudio, m) {}
 #endif
+}
