+++ torchaudio/csrc/sox/utils.h
@@ -15,8 +15,6 @@ namespace sox_utils {
 // APIs for Python interaction
 ////////////////////////////////////////////////////////////////////////////////
 
-bool is_sox_available();
-
 /// Set sox global options
 void set_seed(const int64_t seed);
