+++ torchaudio/utils/sox_utils.py
@@ -2,26 +2,10 @@ from typing import List, Dict
 from functools import wraps
 
 import torch
+from torchaudio._internal import module_utils as _mod_utils
 
 
-def is_sox_available():
-    return torch.ops.torchaudio.is_sox_available()
-
-
-def requires_sox():
-    if is_sox_available():
-        def decorator(func):
-            return func
-    else:
-        def decorator(func):
-            @wraps(func)
-            def wrapped(*args, **kwargs):
-                raise RuntimeError(f'{func.__module__}.{func.__name__} requires sox')
-            return wrapped
-    return decorator
-
-
-@requires_sox()
+@_mod_utils.requires_sox()
 def set_seed(seed: int):
     """Set libsox's PRNG
 
@@ -34,7 +18,7 @@ def set_seed(seed: int):
     torch.ops.torchaudio.sox_utils_set_seed(seed)
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def set_verbosity(verbosity: int):
     """Set libsox's verbosity
 
@@ -52,7 +36,7 @@ def set_verbosity(verbosity: int):
     torch.ops.torchaudio.sox_utils_set_verbosity(verbosity)
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def set_buffer_size(buffer_size: int):
     """Set buffer size for sox effect chain
 
@@ -65,7 +49,7 @@ def set_buffer_size(buffer_size: int):
     torch.ops.torchaudio.sox_utils_set_buffer_size(buffer_size)
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def set_use_threads(use_threads: bool):
     """Set multithread option for sox effect chain
 
@@ -79,7 +63,7 @@ def set_use_threads(use_threads: bool):
     torch.ops.torchaudio.sox_utils_set_use_threads(use_threads)
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def list_effects() -> Dict[str, str]:
     """List the available sox effect names
 
@@ -89,7 +73,7 @@ def list_effects() -> Dict[str, str]:
     return dict(torch.ops.torchaudio.sox_utils_list_effects())
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def list_read_formats() -> List[str]:
     """List the supported audio formats for read
 
@@ -99,7 +83,7 @@ def list_read_formats() -> List[str]:
     return torch.ops.torchaudio.sox_utils_list_read_formats()
 
 
-@requires_sox()
+@_mod_utils.requires_sox()
 def list_write_formats() -> List[str]:
     """List the supported audio formats for write
